+++ services/ProductService.ts
@@ -1,19 +1,20 @@
+import { Product } from './../types/Product.interface'
 import apiClient from '~/services/AxiosConfig'
 
 export default {
-  getProducts() {
+  getProducts(): Promise<any> {
     return apiClient.get(`/products?is_deleted=false`)
   },
-  getProduct(id) {
+  getProduct(id: string): Promise<any> {
     return apiClient.get(`products/${id}`)
   },
-  addProduct(product) {
+  addProduct(product: Product): Promise<any> {
     return apiClient.post('/products', product)
   },
-  updateProduct(product) {
+  updateProduct(product: Product): Promise<any> {
     return apiClient.put(`products/${product.id}`, product)
   },
-  deleteProduct(product) {
+  deleteProduct(product: Product): Promise<any> {
     return apiClient.put(`products/${product.id}`, product)
   },
 }
